employee Section checked (output) and improved

+++ src/main/java/com/day9exercise/demo/Start.java
@@ -18,12 +18,23 @@ import java.util.concurrent.atomic.AtomicBoolean;
 
 public class Start {
 
-    public void restart(CustomerController customerController, CountryController countryController, EmployeeController employeeController, FlightController flightController){
+    public void restartEmployeeSection(CustomerController customerController, CountryController countryController, EmployeeController employeeController, FlightController flightController){
         Scanner scanner = new Scanner(System.in);
         System.out.println("Is there anything else you would like to do?");
         String input = scanner.nextLine();
         if(input.toLowerCase().trim().equals("y")){
             employeeSection(customerController, countryController, employeeController, flightController);
+        } else {
+            System.exit(0);
+        }
+    }
+
+    public void restart(CustomerController customerController, CountryController countryController, EmployeeController employeeController, FlightController flightController){
+        Scanner scanner = new Scanner(System.in);
+        System.out.println("Is there anything else you would like to do?");
+        String input = scanner.nextLine();
+        if(input.toLowerCase().trim().equals("y")){
+            start(customerController, countryController, employeeController, flightController);
         } else {
             System.exit(0);
         }
@@ -40,8 +51,8 @@ public class Start {
             int id = scanner.nextInt();
             System.out.println("Please enter your username");
             scanner.nextLine();
-            String username = scanner.nextLine();
-            System.out.println("Please enter your password");
+            String username = scanner.nextLine().toLowerCase().trim();
+            System.out.println("Please enter your password (case sensitive) ");
             String password  = scanner.nextLine();
 
             AtomicBoolean loggingsuccessful = employeeController.checkEmployee(id, username, password);
@@ -68,7 +79,7 @@ public class Start {
     ///////////////////                     EMPLOYEE SECTION                      /////////////////////////
     public void employeeSection(CustomerController customerController, CountryController countryController, EmployeeController employeeController, FlightController flightController){
         Scanner scanner = new Scanner(System.in);
-            System.out.println("logging successfully.\nWhat would you like to do?\n1. Search customer by passport?\n2. Delete customer by passport?\n3. Amend customer details\n4. View full list of customers\n5. Add new country destination\n6. Delete country from list\n7. Update country\n8. Find country details by name\n9. Add new employee\n10. Employees full list\n11. Update current employees\n");
+            System.out.println("\nWhat would you like to do?\n1. Search customer by passport?\n2. Delete customer by passport?\n3. Amend customer details\n4. View full list of customers\n5. Add new country destination\n6. Delete country from list\n7. Update country\n8. Find country details by country id\n9. Add new employee\n10. Employees full list\n11. Update current employees\n");
             int userInput = scanner.nextInt();
         switch (userInput){
             case 1:
@@ -80,8 +91,14 @@ public class Start {
             case 2:
                 System.out.println("Please enter the customer passport number");
                 scanner.nextLine();
-                String customerPassportDelete = scanner.nextLine();
-                customerController.deleteCustomer(customerPassportDelete);
+                String customerPassportDelete = scanner.nextLine().trim();
+                Optional<Customer> deleteCustomer = customerController.requestedCustomer(customerPassportDelete);
+                System.out.println("Are you sure you want to delete: " + deleteCustomer);
+                String customerInput = scanner.nextLine().toLowerCase().trim();
+                if(customerInput.equals("y")){
+                    customerController.deleteCustomer(customerPassportDelete);
+                }
+                restartEmployeeSection(customerController, countryController, employeeController, flightController);
                 break;
             case 3:
                 System.out.println("Please type in the customer id number");
@@ -89,9 +106,11 @@ public class Start {
                 System.out.println("Please type the required customer information that needs to be updated.\n1. first name\n2. surname\n3. passport number");
                 int updateInfo = scanner.nextInt();
                 customerController.updateCustomer(customerId, updateInfo);
+                restartEmployeeSection(customerController, countryController, employeeController, flightController);
                 break;
             case 4:
                 customerController.listAllCustomers();
+                restartEmployeeSection(customerController, countryController, employeeController, flightController);
                 break;
             case 5:
                 System.out.println("You would like to add a new country destination");
@@ -111,52 +130,43 @@ public class Start {
                 String flightNumber = scanner.nextLine();
                 Country newCountry = new Country(countryName, estimatedDuration, price, arrivalDateTime, departureDateTime, flightNumber);
                 countryController.insertNewCountry(newCountry);
-                restart(customerController, countryController, employeeController, flightController);
+                restartEmployeeSection(customerController, countryController, employeeController, flightController);
                 break;
             case 6:
                 System.out.println("Please enter the country id you want to delete from the system");
                 int countryId = scanner.nextInt();
                 countryController.deleteCountry(countryId);
+                restartEmployeeSection(customerController, countryController, employeeController, flightController);
                 break;
             case 7:
                 System.out.println("Please enter the country id you want to update");
                 int updateCountry = scanner.nextInt();
-                System.out.println("which of the following do you need to update?\n1. Country name?\n2. Estimated Travel?\n3.Price");
-                int requestedUpdate = scanner.nextInt();
-                countryController.updateCountry(updateCountry, requestedUpdate);
+                System.out.println("Is this the specific country you want to update?\n" + countryController.requestedCountry(updateCountry));
+                scanner.nextLine();
+                String userAnswer = scanner.nextLine().toLowerCase().trim();
+                if(userAnswer.equals("y")){
+                    System.out.println("which of the following do you need to update?\n1. Country name?\n2. Estimated Travel?\n3. Price\n4. Flight Number\n5. Time and Date departure\n");
+                    int requestedUpdate = scanner.nextInt();
+                    countryController.updateCountry(updateCountry, requestedUpdate);
+                    restartEmployeeSection(customerController, countryController, employeeController, flightController);
+                } else {
+                    restartEmployeeSection(customerController, countryController, employeeController, flightController);
+                }
                 break;
             case 8:
                 System.out.println("Please enter the country id");
                 scanner.nextLine();
                 int requestedCountryId = scanner.nextInt();
                 countryController.requestedCountry(requestedCountryId);
+                restartEmployeeSection(customerController, countryController, employeeController, flightController);
                 break;
             case 9:
-                System.out.println("Please enter your username");
-                scanner.nextLine();
-                String newUsername = scanner.nextLine();
-                System.out.println("Please enter your password");
-                String newPassword = scanner.nextLine();
-                System.out.println("Please enter your first name");
-                String newFirstName = scanner.nextLine();
-                System.out.println("Please enter your surname");
-                String newSurname = scanner.nextLine();
-                System.out.println("Please enter your National Insurance number");
-                String newNationalInsurance = scanner.nextLine();
-                System.out.println("Are you a current employee? y/n");
-                String answer = scanner.nextLine();
-                Boolean currentEmployee = false;
-                if(answer.toLowerCase().trim().equals("y")){
-                    currentEmployee = true;
-                } else {
-                    currentEmployee = false;
-                }
-                Employee newEmployee = new Employee(newUsername, newPassword, newFirstName, newSurname, newNationalInsurance, currentEmployee);
-                employeeController.insertNewEmployee(newEmployee);
-                System.out.println("Thank you, your employee id is " + newEmployee.getId());
+                registeringEmployee(employeeController);
+                restartEmployeeSection(customerController, countryController, employeeController, flightController);
                 break;
             case 10:
                 employeeController.employeesFullList();
+                restartEmployeeSection(customerController, countryController, employeeController, flightController);
                 break;
             case 11:
                 System.out.println("Please enter your employee id");
@@ -164,6 +174,7 @@ public class Start {
                 System.out.println("Please enter whether you want to update\n1. Username\n2. Password\n3. First name\n4. Surname\n5. Current Employee");
                 int update = scanner.nextInt();
                 employeeController.updateCurrentEmployees(employeeId, update);
+                restartEmployeeSection(customerController, countryController, employeeController, flightController);
                 break;
             default:
                 employeeSection(customerController, countryController, employeeController, flightController);
@@ -276,6 +287,38 @@ public class Start {
         } else {
             Flight newFlight = new Flight(countryId, customerId, returnFlights, numberOfPassenger, totalPrice, customerFlightNumber);
             flightController.addNewFlight(newFlight);
+        }
+    }
+
+    public void registeringEmployee(EmployeeController employeeController){
+        Scanner scanner = new Scanner(System.in);
+        System.out.println("Please enter your username (new user)");
+        String newUsername = scanner.nextLine();
+        System.out.println("Please create your password (new user)");
+        String newPassword = scanner.nextLine();
+        System.out.println("Please confirm your password (new user)");
+        String passwordConfirmed = scanner.nextLine();
+        if(newPassword.equals(passwordConfirmed)){
+            System.out.println("Please enter your first name");
+            String newFirstName = scanner.nextLine();
+            System.out.println("Please enter your surname");
+            String newSurname = scanner.nextLine();
+            System.out.println("Please enter your National Insurance number");
+            String newNationalInsurance = scanner.nextLine();
+            System.out.println("Are you a current employee? y/n");
+            String answer = scanner.nextLine();
+            Boolean currentEmployee = false;
+            if(answer.toLowerCase().trim().equals("y")){
+                currentEmployee = true;
+                } else {
+                currentEmployee = false;
+            }
+            Employee newEmployee = new Employee(newUsername, newPassword, newFirstName, newSurname, newNationalInsurance, currentEmployee);
+            employeeController.insertNewEmployee(newEmployee);
+            System.out.println("Thank you, your employee id is " + newEmployee.getId());
+        } else {
+            System.out.println("Your password does not match.");
+            registeringEmployee(employeeController);
         }
     }
